requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
email-validator
python-multipart
python-dotenv
httpx